requirements.txt
openai>=1.0
python-dotenv>=1.0
sentence-transformers>=2.2
scikit-learn>=1.3
textstat>=0.7
rich>=13.0
pytest>=7.0
pytest-cov>=4.0
celery>=5.3
redis>=5.0
requests>=2.31
anthropic>=0.18
google-generativeai>=0.4

